// src/types/types.ts
export type CreatureKey =
  | 'Pikeman'
  | 'Halberdier'
  | 'Archer'
  | 'Marksman'
  | 'Griffin'
  | 'RoyalGriffin'
  | 'Swordsman'
  | 'Crusader'
  | 'Monk'
  | 'Zealot'
  | 'Cavalier'
  | 'Champion'
  | 'Angel'
  | 'Archangel'
  | 'Centaur'
  | 'CentaurCaptain'
  | 'Dwarf'
  | 'BattleDwarf'
  | 'WoodElf'
  | 'GrandElf'
  | 'Pegasus'
  | 'SilverPegasus'
  | 'DendroidGuard'
  | 'DendroidSoldier'
  | 'Unicorn'
  | 'WarUnicorn'
  | 'GreenDragon'
  | 'GoldDragon'
  | 'Gremlin'
  | 'MasterGremlin'
  | 'StoneGargoyle'
  | 'ObsidianGargoyle'
  | 'StoneGolem'
  | 'IronGolem'
  | 'Mage'
  | 'ArchMage'
  | 'Genie'
  | 'MasterGenie'
  | 'Naga'
  | 'NagaQueen'
  | 'Giant'
  | 'Titan'
  | 'Skeleton'
  | 'SkeletonWarrior'
  | 'WalkingDead'
  | 'Zombie'
  | 'Wight'
  | 'Wraith'
  | 'Vampire'
  | 'VampireLord'
  | 'Lich'
  | 'PowerLich'
  | 'BlackKnight'
  | 'DreadKnight'
  | 'BoneDragon'
  | 'GhostDragon'
  | 'Troglodyte'
  | 'InfernalTroglodyte'
  | 'Harpy'
  | 'HarpyHag'
  | 'Beholder'
  | 'EvilEye'
  | 'Medusa'
  | 'MedusaQueen'
  | 'Minotaur'
  | 'MinotaurKing'
  | 'Manticore'
  | 'Scorpicore'
  | 'RedDragon'
  | 'BlackDragon'
  | 'Imp'
  | 'Familiar'
  | 'Gog'
  | 'Magog'
  | 'HellHound'
  | 'Cerberus'
  | 'Demon'
  | 'HornedDemon'
  | 'PitFiend'
  | 'PitLord'
  | 'Efreeti'
  | 'EfreetSultan'
  | 'Devil'
  | 'ArchDevil'
  | 'Gnoll'
  | 'GnollMarauder'
  | 'Lizardman'
  | 'LizardWarrior'
  | 'SerpentFly'
  | 'DragonFly'
  | 'Basilisk'
  | 'GreaterBasilisk'
  | 'Gorgon'
  | 'MightyGorgon'
  | 'Wyvern'
  | 'WyvernMonarch'
  | 'Hydra'
  | 'ChaosHydra'
  | 'Goblin'
  | 'Hobgoblin'
  | 'WolfRider'
  | 'WolfRaider'
  | 'Orc'
  | 'OrcChieftain'
  | 'Ogre'
  | 'OgreMage'
  | 'Roc'
  | 'Thunderbird'
  | 'Cyclops'
  | 'CyclopsKing'
  | 'Behemoth'
  | 'AncientBehemoth'
  | 'Pixie'
  | 'Sprite'
  | 'AirElemental'
  | 'StormElemental'
  | 'WaterElemental'
  | 'IceElemental'
  | 'FireElemental'
  | 'EnergyElemental'
  | 'EarthElemental'
  | 'MagmaElemental'
  | 'PsychicElemental'
  | 'MagicElemental'
  | 'Firebird'
  | 'Phoenix'
  | 'Peasant'
  | 'Halfling'
  | 'Boar'
  | 'Rogue'
  | 'Nomad'
  | 'Mummy'
  | 'Sharpshooter'
  | 'Troll'
  | 'GoldGolem'
  | 'DiamondGolem'
  | 'Enchanter'
  | 'FaerieDragon'
  | 'RustDragon'
  | 'CrystalDragon'
  | 'AzureDragon'
  | 'Ballista'
  | 'Nymph'
  | 'Oceanid'
  | 'CrewMate'
  | 'Seaman'
  | 'Pirate'
  | 'Corsair'
  | 'SeaDog'
  | 'Stormbird'
  | 'Ayssid'
  | 'SeaWitch'
  | 'Sorceress'
  | 'Nix'
  | 'NixWarrior'
  | 'SeaSerpent'
  | 'Haspid'
  | 'Fangarm'
  | 'Satyr'
  | 'Leprechaun'
  | 'Cannon'
  | 'SteelGolem'
// ...
export type SpellKey =
  | 'MagicArrow'
  | 'Visions'
  | 'Bloodlust'
  | 'Curse'
  | 'ProtectionFromFire'
  | 'Blind'
  | 'FireWall'
  | 'Fireball'
  | 'LandMine'
  | 'Misfortune'
  | 'Armageddon'
  | 'Berserk'
  | 'FireShield'
  | 'Frenzy'
  | 'Inferno'
  | 'Slayer'
  | 'Sacrifice'
  | 'SummonFireElemental'
  | 'Bless'
  | 'Cure'
  | 'Dispel'
  | 'ProtectionFromWater'
  | 'SummonBoat'
  | 'IceBolt'
  | 'RemoveObstacle'
  | 'ScuttleBoat'
  | 'Weakness'
  | 'Forgetfulness'
  | 'FrostRing'
  | 'Mirth'
  | 'Teleport'
  | 'Clone'
  | 'Prayer'
  | 'WaterWalk'
  | 'SummonWaterElemental'
  | 'Shield'
  | 'Slow'
  | 'StoneSkin'
  | 'ViewEarth'
  | 'Quicksand'
  | 'AnimateDead'
  | 'AntiMagic'
  | 'Earthquake'
  | 'ForceField'
  | 'ProtectionFromEarth'
  | 'MeteorShower'
  | 'Resurrection'
  | 'Sorrow'
  | 'TownPortal'
  | 'Implosion'
  | 'SummonEarthElemental'
  | 'Haste'
  | 'ViewAir'
  | 'DeathRipple'
  | 'Disguise'
  | 'DisruptingRay'
  | 'Fortune'
  | 'LightningBolt'
  | 'Precision'
  | 'ProtectionFromAir'
  | 'AirShield'
  | 'DestroyUndead'
  | 'Hypnotize'
  | 'ChainLightning'
  | 'Counterstrike'
  | 'DimensionDoor'
  | 'Fly'
  | 'MagicMirror'
  | 'SummonAirElemental'
  | 'TitansLightningBolt'

export type TerrainKey =
  | 'Neutral'
  | 'Grass'
  | 'Snow'
  | 'Lava'
  | 'Dirt'
  | 'Subterranean'
  | 'Rough'
  | 'Swamp'
  | 'Highlands'
  | 'Sand'
  | 'Wasteland'

export type TownKey =
  | 'Castle'
  | 'Rampart'
  | 'Tower'
  | 'Inferno'
  | 'Necropolis'
  | 'Dungeon'
  | 'Stronghold'
  | 'Fortress'
  | 'Conflux'
  | 'Cove'
  | 'Factory'

export type BattleSide = 'attacker' | 'defender'